backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
langchain==0.0.267
faiss-cpu==1.7.4
openai==0.28.0
tiktoken==0.4.0
google-generativeai==0.3.1
